Removed unnecessary code
